Group plotting tests
