+++ tests/baseline_images/test_image_plots/compare_with_other.png
