+++ tests/baseline_images/test_image_plots/plane_show.png
