+++ tests/baseline_images/test_image_plots/show_image.png
