+++ tests/baseline_images/test_mesh_plots/compare_mesh_with_image.png
