+++ tests/baseline_images/test_mesh_plots/contour_plot.png
