+++ tests/baseline_images/test_mesh_plots/line_mesh.png
