+++ tests/baseline_images/test_mesh_plots/line_triangle_plot.png
