+++ tests/baseline_images/test_mesh_plots/triangle_mesh.png
