+++ tests/segmented_image.npy
